More update

+++ src/object/appear/bullet/NormalBullet.java
@@ -7,7 +7,7 @@ import essential.ZIndex;
 
 public class NormalBullet extends Bullet {
 	
-	public NormalBullet(IAttackable parent, int damage, int x, int y, int angle) {
+	public NormalBullet(IAttackable parent, int damage, int x, int y, float angle) {
 		super(damage, x, y, angle, 2, 5);
 		this.parent = parent;
 		render = new CircleRendable(x, y, 5, colBullet, ZIndex.BULLET);
